+++ test/fuzz/Liquidator/SetAuctionCurveParameters.fuzz.t.sol
@@ -25,8 +25,8 @@ contract SetAuctionCurveParameters_Liquidator_Fuzz_Test is Liquidator_Fuzz_Test 
     //////////////////////////////////////////////////////////////*/
     function testFuzz_Revert_setAuctionCurveParameters_NonOwner(
         address unprivilegedAddress_,
-        uint16 halfLifeTime,
-        uint16 cutoffTime
+        uint32 halfLifeTime,
+        uint32 cutoffTime
     ) public {
         vm.assume(unprivilegedAddress_ != users.creatorAddress);
 
@@ -36,7 +36,7 @@ contract SetAuctionCurveParameters_Liquidator_Fuzz_Test is Liquidator_Fuzz_Test 
         vm.stopPrank();
     }
 
-    function testFuzz_Revert_setAuctionCurveParameters_BaseTooHigh(uint16 halfLifeTime, uint16 cutoffTime) public {
+    function testFuzz_Revert_setAuctionCurveParameters_BaseTooHigh(uint32 halfLifeTime, uint32 cutoffTime) public {
         // Preprocess: limit the fuzzing to acceptable levels
         vm.assume(halfLifeTime > 8 * 60 * 60);
 
@@ -47,7 +47,7 @@ contract SetAuctionCurveParameters_Liquidator_Fuzz_Test is Liquidator_Fuzz_Test 
         vm.stopPrank();
     }
 
-    function testFuzz_Revert_setAuctionCurveParameters_BaseTooLow(uint16 halfLifeTime, uint16 cutoffTime) public {
+    function testFuzz_Revert_setAuctionCurveParameters_BaseTooLow(uint32 halfLifeTime, uint32 cutoffTime) public {
         // Preprocess: limit the fuzzing to acceptable levels
         vm.assume(halfLifeTime < 2 * 60);
 
@@ -58,7 +58,7 @@ contract SetAuctionCurveParameters_Liquidator_Fuzz_Test is Liquidator_Fuzz_Test 
         vm.stopPrank();
     }
 
-    function testFuzz_Revert_setAuctionCurveParameters_AuctionCutoffTimeTooHigh(uint16 halfLifeTime, uint16 cutoffTime)
+    function testFuzz_Revert_setAuctionCurveParameters_AuctionCutoffTimeTooHigh(uint32 halfLifeTime, uint32 cutoffTime)
         public
     {
         // Preprocess: limit the fuzzing to acceptable levels
@@ -74,7 +74,7 @@ contract SetAuctionCurveParameters_Liquidator_Fuzz_Test is Liquidator_Fuzz_Test 
         vm.stopPrank();
     }
 
-    function testFuzz_Revert_setAuctionCurveParameters_AuctionCutoffTimeTooLow(uint16 halfLifeTime, uint16 cutoffTime)
+    function testFuzz_Revert_setAuctionCurveParameters_AuctionCutoffTimeTooLow(uint32 halfLifeTime, uint32 cutoffTime)
         public
     {
         // Preprocess: limit the fuzzing to acceptable levels
@@ -90,7 +90,7 @@ contract SetAuctionCurveParameters_Liquidator_Fuzz_Test is Liquidator_Fuzz_Test 
         vm.stopPrank();
     }
 
-    function testFuzz_Revert_setAuctionCurveParameters_PowerFunctionReverts(uint8 halfLifeTime, uint16 cutoffTime)
+    function testFuzz_Revert_setAuctionCurveParameters_PowerFunctionReverts(uint32 halfLifeTime, uint32 cutoffTime)
         public
     {
         // Preprocess: limit the fuzzing to acceptable levels
@@ -105,7 +105,7 @@ contract SetAuctionCurveParameters_Liquidator_Fuzz_Test is Liquidator_Fuzz_Test 
         vm.stopPrank();
     }
 
-    function testFuzz_Success_setAuctionCurveParameters_Base(uint16 halfLifeTime, uint16 cutoffTime) public {
+    function testFuzz_Success_setAuctionCurveParameters_Base(uint32 halfLifeTime, uint32 cutoffTime) public {
         // Preprocess: limit the fuzzing to acceptable levels
         vm.assume(halfLifeTime > 2 * 60);
         vm.assume(halfLifeTime < 8 * 60 * 60);
@@ -126,7 +126,7 @@ contract SetAuctionCurveParameters_Liquidator_Fuzz_Test is Liquidator_Fuzz_Test 
         assertEq(liquidator.getBase(), expectedBase);
     }
 
-    function testFuzz_Success_setAuctionCurveParameters_cutoffTime(uint16 halfLifeTime, uint16 cutoffTime) public {
+    function testFuzz_Success_setAuctionCurveParameters_cutoffTime(uint32 halfLifeTime, uint32 cutoffTime) public {
         // Preprocess: limit the fuzzing to acceptable levels
         vm.assume(halfLifeTime > 1 * 60 * 60);
         vm.assume(halfLifeTime < 8 * 60 * 60);
